+++ src/Brendt/Stitcher/Application/DevController.php
@@ -2,6 +2,7 @@
 
 namespace Brendt\Stitcher\Application;
 
+use Brendt\Stitcher\App;
 use Brendt\Stitcher\Exception\StitcherException;
 use Brendt\Stitcher\Factory\AdapterFactory;
 use Brendt\Stitcher\Stitcher;
@@ -17,17 +18,10 @@ use Symfony\Component\Routing\RouteCollection;
  */
 class DevController
 {
+    const ENVIRONMENT = 'development';
 
-    /**
-     * @var Stitcher
-     */
     protected $stitcher;
 
-    /**
-     * Create a new developer controller with optional configuration path and -file.
-     *
-     * @param Stitcher $stitcher
-     */
     public function __construct(Stitcher $stitcher) {
         $this->stitcher = $stitcher;
     }
@@ -54,11 +48,6 @@ class DevController
         }
     }
 
-    /**
-     * Create the route container
-     *
-     * @return \Symfony\Component\Routing\RouteCollection
-     */
     protected function createRouteCollection() {
         $routeCollection = new RouteCollection();
         $site = $this->stitcher->loadSite();
@@ -77,11 +66,6 @@ class DevController
         return $routeCollection;
     }
 
-    /**
-     * @param $url
-     *
-     * @return mixed
-     */
     protected function getPage($url) {
         $routeCollection = $this->createRouteCollection();
         $matcher = new UrlMatcher($routeCollection, new RequestContext());
@@ -107,5 +91,4 @@ class DevController
 
         throw new ResourceNotFoundException();
     }
-
 }