+++ src/Brendt/Stitcher/Factory/TemplateEngineFactory.php
@@ -11,34 +11,14 @@ class TemplateEngineFactory
     const SMARTY_ENGINE = 'smarty';
     const TWIG_ENGINE = 'twig';
 
-    /**
-     * @var ContainerInterface
-     */
     private $container;
-
-    /**
-     * @var string
-     */
     private $templateEngine;
 
-    /**
-     * TemplateEngineFactory constructor.
-     *
-     * @param ContainerInterface $container
-     * @param string             $templateEngine
-     */
     public function __construct(ContainerInterface $container, string $templateEngine) {
         $this->container = $container;
         $this->templateEngine = $templateEngine;
     }
 
-    /**
-     * @param $type
-     *
-     * @return mixed
-     *
-     * @throws UnknownEngineException
-     */
     public function getByType($type) : TemplateEngine {
         switch ($type) {
             case self::TWIG_ENGINE:
@@ -50,11 +30,7 @@ class TemplateEngineFactory
         }
     }
 
-    /**
-     * @return TemplateEngine
-     */
     public function getDefault() : TemplateEngine {
         return $this->getByType($this->templateEngine);
     }
-
 }