+++ src/Brendt/Stitcher/Parser/DefaultParser.php
@@ -4,14 +4,6 @@ namespace Brendt\Stitcher\Parser;
 
 class DefaultParser implements Parser
 {
-
-    /**
-     * Parse a path into usable data.
-     *
-     * @param $path
-     *
-     * @return mixed
-     */
     public function parse($path) {
         return $path;
     }