+++ src/Brendt/Stitcher/Parser/Parser.php
@@ -7,14 +7,5 @@ namespace Brendt\Stitcher\Parser;
  */
 interface Parser
 {
-
-    /**
-     * Parse a path into usable data.
-     *
-     * @param $path
-     *
-     * @return mixed
-     */
     public function parse($path);
-
 }