+++ src/Brendt/Stitcher/Parser/Site/PageParser.php
@@ -16,35 +16,11 @@ use Symfony\Component\Finder\SplFileInfo;
 
 class PageParser
 {
-
-    /**
-     * @var ParserFactory
-     */
     private $parserFactory;
-
-    /**
-     * @var MetaCompiler
-     */
     private $metaCompiler;
-
-    /**
-     * @var TemplatePlugin
-     */
     private $templatePlugin;
-
-    /**
-     * @var string
-     */
     private $templateDir;
-
-    /**
-     * @var AdapterFactory
-     */
     private $adapterFactory;
-
-    /**
-     * @var HeaderCompiler|null
-     */
     private $headerCompiler;
 
     public function __construct(
@@ -75,25 +51,21 @@ class PageParser
      * @return SplFileInfo[]
      */
     public function loadTemplates() {
-        $templateExtension = $this->templateEngine->getTemplateExtension();
+        $templateExtensions = $this->templateEngine->getTemplateExtensions();
+        $templateExtensionsRegex = '/\.(' . implode('|', $templateExtensions) . ')/';
 
         /** @var SplFileInfo[] $files */
-        $files = Finder::create()->files()->in($this->templateDir)->name("*.{$templateExtension}");
+        $files = Finder::create()->files()->in($this->templateDir)->name($templateExtensionsRegex);
         $templates = [];
 
         foreach ($files as $file) {
-            $id = str_replace(".{$templateExtension}", '', $file->getRelativePathname());
+            $id = preg_replace($templateExtensionsRegex, '', $file->getRelativePathname());
             $templates[$id] = $file;
         }
 
         return $templates;
     }
 
-    /**
-     * @param Page $page
-     *
-     * @return string
-     */
     public function parsePage(Page $page) : string {
         $entryPage = $this->parseVariables($page);
         $this->metaCompiler->compilePage($page);
@@ -173,11 +145,6 @@ class PageParser
         return $page;
     }
 
-    /**
-     * @param Page $page
-     *
-     * @throws TemplateNotFoundException
-     */
     public function validate(Page $page) {
         $templateIsset = isset($this->templates[$page->getTemplatePath()]);